the last thing you want in your Buger King Buger is someonelses foot fungis, but it that might be what  you get

+++ src/main/java/Student.java
@@ -96,7 +96,7 @@ public class Student extends Person {
     *
     * **/
     public void learn(int classHours){
-        this.credits += classHours/10;
+        this.credits += (double) classHours/10;
 
     }
